+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Enhanced: 📦 Load chosen shipping method when loading cart from session via the web.
 * Enhanced: 📦 Load cart fees when loading cart from session via the web.
 * Dev: **NEW** filter `cocart_load_cart_query_name` to allow developers add more white labelling when loading the cart via the web.
+* Improved: `uninstall.php` to delete WooCommerce Admin notes when uninstalling for those who are using WooCommerce v4.8 or greater.
 
 ## v2.8.1 - 10th December, 2020